Added imperative LangChain implementation

+++ conversation-ai-ws/src/main/java/com/thetealover/conversation/ws/config/properties/AiConfigurationProperties.java
@@ -17,6 +17,6 @@ public interface AiConfigurationProperties {
 
     Double topP();
 
-    Double topK();
+    Integer topK();
   }
 }
